Fixed #18 (remarking a task as completed every time that it loads), but had to so by moving FileParser to the task package—which is not a desired behavior (and inadvertently references #17).
